// frontend/src/lib/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.1"
  }
  public: {
    Tables: {
// ...
      user_contracts: {
        Row: {
          contract_address: string
          created_at: string | null
          id: string
          role: string
          user_address: string
        }
        Insert: {
          contract_address: string
          created_at?: string | null
          id?: string
          role: string
          user_address: string
        }
        Update: {
          contract_address?: string
          created_at?: string | null
          id?: string
          role?: string
          user_address?: string
        }
        Relationships: [
          {
            foreignKeyName: "fk_contract"
            columns: ["contract_address"]
            isOneToOne: false
            referencedRelation: "rental_contracts"
            referencedColumns: ["contract_address"]
          },
          {
            foreignKeyName: "fk_user"
            columns: ["user_address"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["wallet_address"]
          },
        ]
      }
      users: {
        Row: {
          created_at: string | null
          ens_name: string | null
          id: string
          updated_at: string | null
          wallet_address: string
        }
        Insert: {
          created_at?: string | null
          ens_name?: string | null
          id?: string
          updated_at?: string | null
          wallet_address: string
        }
        Update: {
          created_at?: string | null
          ens_name?: string | null
          id?: string
          updated_at?: string | null
          wallet_address?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}